% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthplotr-package.R
\docType{package}
\name{healthplotr-package}
\alias{healthplotr-package}
\alias{_PACKAGE}
\title{healthplotr: Shiny App to Create Plots from Apple Health Data}
\description{
This package creates plots from Apple Health data. It does so via a Shiny web app, where a user can upload a file called export.zip, a zip file generated by Apple's Health app.
}
\author{
\strong{Maintainer}: Mark Moll \email{mark@moll.ai} (\href{https://orcid.org/0000-0002-0451-2797}{ORCID})

}
\keyword{internal}
