% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{healthplotr}
\alias{healthplotr}
\title{Start a Shiny app to visualize Apple Health data}
\usage{
healthplotr(healthdata = NULL)
}
\arguments{
\item{healthdata}{(1) a zip file with Apple Health,
(2) an RDS file with health data as returned by healthplotr::load_exported_health_data, or
(3) a health data object as returned by healthplotr::load_exported_health_data}
}
\description{
Start a Shiny app to visualize Apple Health data
}
