% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdata.R
\name{load_exported_health_data}
\alias{load_exported_health_data}
\title{Load Apple health data from a zip file and return a list of 3 tidy tibbles
with records, activity, and workout data}
\usage{
load_exported_health_data(zip_file)
}
\arguments{
\item{zip_file}{path to a zip file created by the iOS Health app.}
}
\description{
Load Apple health data from a zip file and return a list of 3 tidy tibbles
with records, activity, and workout data
}
