% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{plot_activity}
\alias{plot_activity}
\title{Generate a plot of a Apple Health activity type}
\usage{
plot_activity(data, type, aggregate)
}
\arguments{
\item{data}{A tibble with Apple Health activity data}

\item{type}{A string name of the activity type ("activeEnergyBurned", "appleExerciseTime", or "appleStandHours")}

\item{aggregate}{A time period string ("year", "month", "week", "day")}
}
\value{
A plotly object
}
\description{
Generate a plot of a Apple Health activity type
}
