% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{plot_records}
\alias{plot_records}
\title{Generate a plot of a Apple Health records}
\usage{
plot_records(data, type, aggregate)
}
\arguments{
\item{data}{A tibble with Apple Health records}

\item{type}{A string name of the record type (e.g., "BodyMass")}

\item{aggregate}{A time period string ("year", "month", "week", "day")}
}
\value{
A plotly object
}
\description{
Generate a plot of a Apple Health records
}
