% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout.R
\name{plot_workout}
\alias{plot_workout}
\title{Generate a plot of a Apple Health workout data}
\usage{
plot_workout(data, aggregate)
}
\arguments{
\item{data}{A tibble with Apple Health workout data}

\item{aggregate}{A time period string ("year", "month", "week", "day")}
}
\value{
A ggplot object
}
\description{
Generate a plot of a Apple Health workout data
}
